web-app/requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
Django==2.2.15
django-allauth==0.42.0
django-braces==1.14.0
django-environ==0.4.5
django-filter==2.3.0
django-oauth-toolkit==1.3.2
django-redis==4.12.1
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.1
idna==2.10
kombu==4.6.11
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
django-safedelete==0.5.6
Pillow==7.2.0
django-cors-headers==3.5.0